mshop/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.10.8
django-allauth==0.34.0
django-cart==1.0.4
django-filer==1.2.8
django-filters==0.2.1
django-mptt==0.8.7
django-paypal==0.4.1
django-polymorphic==1.0.2
djangorestframework==3.7.1
easy-thumbnails==2.3
idna==2.6
oauthlib==2.0.6
olefile==0.44
paypal==1.2.5
Pillow==4.3.0
python-openid==2.2.5
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
